Refactor open password for sync to reduce risk of circular dependency
